+++ src/ocean/util/serialize/contiguous/VersionDecorator.d
@@ -42,6 +42,7 @@ import ocean.text.convert.Format,
 
 *******************************************************************************/
 
+deprecated("Use MultiVersionDecorator instead")
 class VersionDecorator
 {
     /***************************************************************************
@@ -52,14 +53,6 @@ class VersionDecorator
 
     public alias VersionDecorator This;
 
-    /**************************************************************************
-
-        NB! This will suppress any compilation errors, comment out during
-        development and enable only when commiting.
-
-    **************************************************************************/
-
-    static assert (isDecorator!(This));
 
     /***************************************************************************
 
@@ -146,7 +139,7 @@ import core.memory;
 
 *******************************************************************************/
 
-unittest
+deprecated unittest
 {
     struct S
     {
@@ -180,7 +173,7 @@ unittest
 
 *******************************************************************************/
 
-unittest
+deprecated unittest
 {
     struct S
     {
@@ -206,7 +199,7 @@ unittest
 
 *******************************************************************************/
 
-unittest
+deprecated unittest
 {
     auto loader = new VersionDecorator;
     void[] buffer = null;
@@ -267,7 +260,7 @@ struct Test1
     }
 }
 
-unittest
+deprecated unittest
 {
     auto loader = new VersionDecorator;
 
@@ -316,7 +309,7 @@ struct Test2
     }
 }
 
-unittest
+deprecated unittest
 {
     auto loader = new VersionDecorator;
 
@@ -560,7 +553,7 @@ struct Test3
     }
 }
 
-Dst testConv(Src, Dst)(Src src)
+deprecated Dst testConv(Src, Dst)(Src src)
 {
     auto test = new NamedTest(Src.stringof ~ " -> " ~ Dst.stringof);
 
@@ -593,7 +586,7 @@ Dst testConv(Src, Dst)(Src src)
     }
 }
 
-unittest
+deprecated unittest
 {
     with (Test3)
     {
@@ -613,7 +606,7 @@ unittest
     }
 }
 
-Dst testConvMemory(Src, Dst)(Src src)
+deprecated Dst testConvMemory(Src, Dst)(Src src)
 {
     auto test = new NamedTest(Src.stringof ~ " -> " ~ Dst.stringof);
 
@@ -669,7 +662,7 @@ Dst testConvMemory(Src, Dst)(Src src)
     return *result.ptr;
 }
 
-unittest
+deprecated unittest
 {
     with (Test3)
     {
@@ -713,7 +706,7 @@ struct Test4
     }
 }
 
-unittest
+deprecated unittest
 {
     auto loader = new VersionDecorator;
     void[] buffer;